requirements.txt
asgiref==3.8.1
Django==4.2.16
django-cors-headers==4.6.0
djangorestframework==3.15.2
psycopg2-binary==2.9.10
python-dotenv==1.0.1
rest-framework-simplejwt==0.0.2
sqlparse==0.5.2
